left-sidebar: Simplify and make more uniform the "All streams" header.
The issue before was that the left sidebar would become too tall for
the screen because the standard header that has “STREAMS” and buttons
is 20px tall, and this one is 30px tall. This makes it much shorter,
changes the text to be the same style as the “STREAMS” text (medium
grey, uppercase text).

The markup is then fixed to be significantly less verbose than before —
changing a list to just a simple link.